Create Action.java
Action is the basic item in the Schedule. It is also the item that the user creates themselves by adding a name and a time in which the action should be completed as well as a duration.

Shifting factor, while not yet implemented, should be used to take schedules from the lists of schedules according to what level of variety in times the user wants for this action.