Create Schedule.java
Schedule is a list of Actions. This list of actions are added to the Schedule, and then when the user has added all items, the method makeScheduleList should be invoked. This creates a set of empty lists, which it passes to threads (in ThreadManager) who then try to create valid non conflicting Schedules.