Schedule Object
The Schedule object allows for lists to be created, added to, checked on and retrieved. It also calls and creates threads to complete the work to find valid schedules. A valid schedule is when the list of Action objects it contains have windows the size of their duration and no two windows of those actions overlap.