GetCombinations explained
Get Combinations is an important algorithm to how the application works. Given the huge number of possible combinations, it is important that we dont have to store them, that we can instead, store something smaller, such as a long, that can be then used to determine which combination it represents.